+++ logistic_regression/Parameters/B.csv
@@ -1,0 +1,50 @@
+-3.963400672665901570e-01
+-3.433360790793736350e-01
+-3.355664067283206653e-01
+-3.317870022267022101e-01
+-3.318192077399519935e-01
+-3.776716209412434577e-01
+-3.628210769449449757e-01
+-3.449875647352558805e-01
+-3.567239034407481690e-01
+-3.591953641326132152e-01
+-3.480885514611739007e-01
+-3.556509562559136972e-01
+-3.656821722906778982e-01
+-3.269567483074835579e-01
+-3.242691087599899880e-01
+-3.599525044766259119e-01
+-3.373575596297205825e-01
+-2.905640431829912274e-01
+-3.264654516922979588e-01
+-3.623946528730968009e-01
+-3.326922314517763990e-01
+-3.399971801674171457e-01
+-3.402542040706037496e-01
+-4.021380985505679240e-01
+-4.052593289751030281e-01
+-3.567337309140132806e-01
+-3.604351364295108051e-01
+-3.833101445489791592e-01
+-3.420653811114624521e-01
+-3.481819095455094826e-01
+-3.001211994047533471e-01
+-3.680697098249952615e-01
+-3.639459624642267244e-01
+-3.403418424264261777e-01
+-3.935524234053360826e-01
+-3.482053218566873443e-01
+-3.802673476062242153e-01
+-3.161607286470528577e-01
+-3.460595916178518805e-01
+-3.446102849528526191e-01
+-3.222971326179743801e-01
+-3.244536848207678625e-01
+-3.255807346539705271e-01
+-3.045248938611994083e-01
+-3.146735944449985389e-01
+-3.240267802643956863e-01
+-3.632574287910459443e-01
+-3.783622177805830433e-01
+-3.669036198442756080e-01
+-2.863963032513943130e-01
