+++ logistic_regression/Parameters/W.csv
@@ -1,0 +1,4 @@
+1.417503858247157522e+00,1.199379416695875777e+00,1.424443073480743172e+00,1.547263034765701262e+00,1.357941969425502382e+00,1.428155141115006410e+00,1.399643488978589589e+00,1.316635921074690518e+00,1.432728442556505000e+00,1.319416207717461154e+00,1.320534102286470279e+00,1.246908147729048011e+00,1.614333102187446300e+00,1.180747737614042059e+00,1.777183097846413595e+00,1.513974383031640336e+00,1.195461009033956579e+00,9.753184400367145290e-01,1.086684503253641410e+00,1.404919636205469891e+00,1.334808239403428187e+00,1.313487541814086557e+00,1.319952273460127223e+00,1.797924810562251841e+00,1.747786360826697916e+00,1.404366031781037361e+00,1.612478956541288699e+00,1.644152334751971090e+00,1.246109889757027522e+00,1.317736996896088142e+00,9.462606660968251315e-01,1.553825284885596902e+00,1.464565232605721246e+00,1.456658055029398158e+00,1.496303875776864079e+00,1.218964530345459485e+00,1.319307808472778731e+00,1.206852448715085346e+00,1.319869999315300690e+00,1.244009488458817092e+00,1.137294380525794901e+00,1.256828621302861215e+00,1.363252185381744752e+00,1.036638074715005908e+00,1.146315813425382135e+00,1.354775892370272405e+00,1.492399245809536312e+00,1.731842934704697301e+00,1.277235303784960863e+00,1.049704934496372255e+00
+1.047096735520995983e-02,2.744720170219556998e-02,3.995627881680791776e-02,3.195627749182070043e-02,3.572156660159928748e-02,3.450154934315766142e-02,3.116587769508163963e-02,3.658613703964389702e-02,3.499247158273666169e-02,4.876077455834441909e-02,4.420890728181566670e-02,3.607556435141741347e-02,3.410448497239855492e-02,6.291538590037308587e-02,4.047584429870543926e-02,4.302687471024101845e-02,5.069982200044605147e-02,5.531139601793546440e-02,4.237834288170693603e-02,4.051645965458272602e-02,2.964844303296911357e-02,3.012910675785608963e-02,4.135553702396504300e-02,3.766917630472558370e-02,4.557517692325223357e-02,3.617803875240969275e-02,3.790751524897133240e-02,4.206573635466100136e-02,2.686973776367613392e-02,3.249925984204463103e-02,4.923562762276646049e-02,2.955187759582057760e-02,4.735211752278469660e-02,4.024080194383070075e-02,4.149427833391090836e-02,4.662139618576629368e-02,4.080803513897202778e-02,4.684903146804382523e-02,4.863254003616604426e-02,3.424208205138958711e-02,2.792211917663973281e-02,5.344066847235043327e-02,4.287413667821979302e-02,1.586163285973308384e-02,2.635572951800523381e-02,4.272549560100508126e-02,5.923362369577640091e-02,2.591078843920615118e-02,3.242456272448159676e-02,2.864220800335412140e-02
+7.093067175955725112e-02,3.985255443868362946e-02,4.171285498763230293e-02,4.273475985148619122e-02,1.844949995753580527e-02,3.910974538515292476e-02,4.110531551819914670e-02,2.903992110443138347e-02,3.212212805376752495e-02,3.038418012669209106e-02,2.008220896511948503e-02,3.768547293114094671e-02,4.761237393971867737e-02,1.872057126654506215e-02,3.697579832994481869e-02,3.222088742008538675e-02,2.450792345072661318e-02,4.405132397890828416e-02,3.477014916335602335e-02,3.979140946989238176e-02,2.514531803305275134e-02,2.962923443655760339e-02,3.485241606172263112e-02,3.978094836900927800e-02,5.359055397973549323e-02,3.494919313088142021e-02,2.870227548418748456e-02,2.655053197664244968e-02,5.699872346035252807e-02,2.589522383637903394e-02,1.803714605697077575e-02,3.579741603605578887e-02,3.948886857361968983e-02,2.336907357291047663e-02,2.732333365928698785e-02,2.875064914100177935e-02,2.741697921174653108e-02,2.394201619566370745e-02,3.194816355982982997e-02,3.985995199451244225e-02,4.023476863882254756e-02,1.489302369456336345e-02,1.511874631109239378e-02,4.980306135705055148e-02,4.040644134906053175e-02,3.058868527443655075e-02,2.577433576847021346e-02,4.184741358772542386e-02,4.671488741836550551e-02,4.372493835060304451e-02
+-7.181660647346273241e-02,-7.058724012039110929e-02,-7.629132139597498419e-02,-3.097176332365097345e-02,-7.132541101003241069e-02,-9.023093765781725617e-02,-6.848748277096755599e-02,-5.978334286999185260e-02,-6.748892534885503336e-02,-2.364612838453292523e-02,-9.654257558741913270e-02,-5.484990900201866526e-02,-5.970014626894169091e-02,-5.012501723802812587e-02,-4.136717157379939114e-02,-3.396117813170166505e-02,-2.874847141161510425e-02,-3.206727927926317223e-02,-6.981191706211087311e-02,-4.498740537584625110e-02,1.329630551420242675e-02,-4.058005678932821020e-02,-5.969452811307105616e-02,-7.142324382627691925e-02,-5.526167197213043697e-02,-4.431933204558267253e-02,-4.697372237537975059e-02,-3.900234851557501936e-02,-4.308921822134444829e-02,-3.654942500943694078e-02,-8.000075203805481761e-02,-7.493766983356144051e-02,-2.943481202945526579e-02,-4.468116610513364617e-02,-7.643726587477744916e-02,-6.141478945230053471e-02,-7.739735720364988858e-02,-7.245988669422193107e-02,-2.305121011751323432e-02,-8.114673629395678656e-02,-6.138038847064640480e-02,-2.022392540377516632e-02,-5.534085391043898350e-02,-3.015836912374803350e-02,-4.846495468616848074e-02,-3.412340451979282985e-02,-5.937764121033763803e-02,-4.095871261704555483e-02,-4.376243617133275532e-02,-4.832196320938039746e-02
