--- logistic_regression/Training_/X_test_HWAY.csv
@@ -1,136 +1,0 @@
-ID,Height,Age,Weight
-2201,180.0,34.0,113.0
-7587,198.0,21.0,106.0
-8238,186.0,26.0,128.0
-9786,183.0,23.0,120.0
-25346,185.0,22.0,93.0
-42303,196.0,24.0,124.0
-51071,186.0,27.0,149.0
-62308,196.0,24.0,125.0
-87674,190.0,32.0,111.0
-92739,192.0,25.0,125.0
-93433,186.0,40.0,107.0
-107613,159.0,26.0,70.0
-110212,188.0,26.0,120.0
-114187,194.0,27.0,122.5
-125791,187.0,24.0,95.0
-129478,194.0,24.0,125.0
-133843,200.0,31.0,124.0
-8761,181.0,28.0,98.0
-28219,176.0,36.0,100.0
-32764,194.0,27.0,116.0
-32957,186.0,30.0,83.0
-39294,185.0,27.0,107.0
-61718,196.0,24.0,127.0
-77425,196.0,25.0,107.0
-92409,184.0,41.0,105.0
-116206,174.0,22.0,83.0
-1856,192.0,27.0,86.0
-30559,200.0,30.0,128.0
-31616,196.0,25.0,102.0
-32297,192.0,32.0,118.0
-32800,180.0,33.0,97.0
-34006,186.0,30.0,114.0
-44591,192.0,28.0,115.0
-44597,192.0,31.0,120.0
-56250,194.0,27.0,114.0
-59781,170.0,26.0,80.0
-62194,200.0,23.0,115.0
-64396,184.0,22.0,80.0
-65694,194.0,31.0,120.0
-71588,192.0,21.0,120.0
-76347,199.0,21.0,104.0
-78659,181.0,29.0,80.0
-84515,192.0,38.0,85.0
-117099,198.0,32.0,120.0
-120053,190.0,29.0,105.0
-126502,192.0,22.0,118.0
-5159,190.0,27.0,125.0
-7927,186.0,26.0,87.0
-7968,195.0,22.0,132.0
-18655,189.0,23.0,127.0
-29666,192.0,23.0,96.0
-40969,193.0,24.0,129.0
-43949,200.0,27.0,128.0
-46317,190.0,22.0,115.0
-49109,182.0,34.0,147.0
-49817,190.0,20.0,82.0
-49910,178.0,36.0,106.0
-50258,185.0,22.0,85.0
-62308,196.0,32.0,125.0
-63392,183.0,27.0,134.0
-68647,190.0,26.0,88.0
-71196,193.0,20.0,118.0
-73758,204.0,26.0,142.0
-76534,200.0,23.0,120.0
-77472,190.0,33.0,120.0
-80287,193.0,26.0,130.0
-85913,190.0,27.0,102.0
-87674,190.0,20.0,111.0
-87674,190.0,24.0,111.0
-102565,195.0,24.0,115.0
-102565,195.0,28.0,115.0
-115290,184.0,25.0,91.0
-115992,191.0,23.0,123.0
-119961,183.0,27.0,89.0
-120385,194.0,25.0,118.0
-130080,180.0,21.0,86.0
-359,190.0,26.0,125.0
-5702,192.0,37.0,120.0
-13127,183.0,32.0,113.0
-22698,183.0,25.0,106.0
-27978,192.0,31.0,115.0
-35717,178.0,32.0,95.0
-46029,185.0,21.0,90.0
-60835,188.0,24.0,115.0
-61238,185.0,31.0,103.0
-63609,188.0,28.0,112.0
-63609,188.0,32.0,112.0
-77576,182.0,32.0,115.0
-77576,182.0,48.0,115.0
-83502,187.0,29.0,96.0
-87848,180.0,22.0,98.0
-95488,186.0,28.0,110.0
-100730,187.0,33.0,112.0
-104098,193.0,23.0,112.0
-113888,192.0,24.0,117.0
-118215,191.0,31.0,120.0
-128266,183.0,23.0,97.0
-135313,190.0,27.0,102.0
-2469,200.0,28.0,130.0
-15669,198.0,26.0,140.0
-17078,181.0,19.0,88.0
-25284,198.0,35.0,125.0
-27323,185.0,22.0,106.0
-28544,188.0,23.0,99.0
-30678,193.0,24.0,110.0
-41788,184.0,34.0,104.0
-44597,192.0,27.0,120.0
-46277,201.0,27.0,126.0
-50258,185.0,26.0,85.0
-53966,182.0,36.0,95.0
-54001,207.0,26.0,120.0
-57280,196.0,29.0,125.0
-57280,196.0,33.0,125.0
-61753,183.0,24.0,95.0
-64326,203.0,25.0,130.0
-79780,195.0,33.0,105.0
-88073,192.0,23.0,125.0
-88073,192.0,28.0,125.0
-88073,192.0,32.0,125.0
-96625,188.0,29.0,110.0
-100682,199.0,29.0,110.0
-100682,199.0,33.0,110.0
-102565,195.0,20.0,115.0
-107643,193.0,28.0,110.0
-107643,193.0,32.0,110.0
-109597,190.0,27.0,88.0
-110584,191.0,34.0,114.0
-117837,191.0,21.0,92.5
-117837,191.0,29.0,92.5
-118220,196.0,31.0,120.0
-121356,193.0,32.0,125.0
-123034,203.0,28.0,120.0
-129356,200.0,22.0,120.0
-130471,193.0,25.0,115.0
-131051,177.0,26.0,90.0
