--- logistic_regression/Training_/X_train_HWAY.csv
@@ -1,253 +1,0 @@
-ID,Height,Age,Weight
-1695,195.0,22.0,120.0
-5191,198.0,24.0,155.0
-5803,188.0,26.0,118.0
-15458,198.0,27.0,150.0
-17923,188.0,24.0,98.0
-19770,194.0,25.0,130.0
-21249,186.0,28.0,125.0
-32496,202.0,31.0,165.0
-37878,191.0,25.0,89.0
-40323,189.0,21.0,118.0
-50016,183.0,27.0,93.0
-61166,196.0,36.0,125.0
-68411,185.0,31.0,100.0
-70432,190.0,30.0,116.0
-72065,191.0,22.0,102.0
-73758,204.0,34.0,142.0
-78020,188.0,25.0,118.0
-79025,196.0,30.0,120.0
-79660,202.0,36.0,140.0
-82976,190.0,27.0,120.0
-86380,190.0,20.0,92.0
-88479,194.0,31.0,125.0
-97119,193.0,37.0,131.0
-109597,190.0,23.0,88.0
-114240,188.0,30.0,108.0
-117837,191.0,29.0,92.5
-122671,185.0,28.0,100.0
-123168,196.0,23.0,140.0
-126095,188.0,32.0,130.0
-126492,187.0,25.0,94.0
-132654,191.0,27.0,130.0
-11145,191.0,23.0,108.0
-11446,177.0,19.0,90.0
-17103,189.0,28.0,120.0
-21546,187.0,22.0,110.0
-21629,185.0,22.0,110.0
-22698,183.0,33.0,106.0
-30152,183.0,46.0,104.0
-31090,178.0,25.0,95.0
-33773,191.0,23.0,115.0
-34592,187.0,22.0,89.0
-37085,180.0,30.0,111.0
-39294,185.0,35.0,107.0
-40373,185.0,28.0,109.0
-40373,185.0,32.0,109.0
-41106,189.0,25.0,115.0
-46953,193.0,27.0,115.0
-52324,187.0,25.0,97.0
-55996,188.0,26.0,95.0
-59290,192.0,22.0,106.0
-61558,194.0,36.0,107.0
-62267,192.0,26.0,115.0
-70341,180.0,25.0,77.0
-83502,187.0,33.0,96.0
-95965,186.0,24.0,105.0
-101611,179.0,29.0,115.0
-103116,190.0,32.0,110.0
-104098,193.0,27.0,112.0
-105132,178.0,23.0,106.0
-110208,185.0,39.0,109.0
-117537,184.0,27.0,104.0
-121464,180.0,23.0,109.0
-123497,178.0,31.0,100.0
-129419,193.0,33.0,120.0
-129419,193.0,37.0,120.0
-2198,186.0,28.0,100.0
-2469,200.0,40.0,130.0
-9379,198.0,28.0,80.0
-11492,194.0,23.0,112.0
-12717,191.0,28.0,115.0
-15754,183.0,24.0,91.0
-21571,194.0,27.0,112.0
-37262,186.0,24.0,90.0
-38731,202.0,24.0,120.0
-38858,185.0,22.0,89.0
-40373,185.0,28.0,109.0
-42127,190.0,33.0,110.0
-50076,194.0,24.0,128.0
-53966,182.0,42.0,95.0
-56182,185.0,22.0,102.0
-61225,178.0,24.0,99.0
-61225,178.0,36.0,99.0
-64687,188.0,29.0,106.0
-70034,176.0,28.0,102.0
-73969,200.0,27.0,145.0
-75708,195.0,29.0,101.0
-75708,195.0,33.0,101.0
-76347,199.0,29.0,104.0
-79979,188.0,37.0,121.0
-83256,207.0,26.0,130.0
-83721,196.0,29.0,123.0
-83721,196.0,41.0,123.0
-84515,192.0,25.0,85.0
-89010,193.0,27.0,122.0
-96625,188.0,25.0,110.0
-97430,194.0,32.0,120.0
-108505,195.0,24.0,111.0
-108528,198.0,26.0,120.0
-112391,197.0,23.0,129.0
-112845,175.0,27.0,110.0
-117225,176.0,33.0,84.0
-117443,189.0,28.0,106.0
-120792,184.0,31.0,85.0
-128842,186.0,34.0,109.0
-131821,190.0,20.0,105.0
-4094,191.0,25.0,118.0
-6810,194.0,29.0,135.0
-7968,195.0,30.0,132.0
-8385,198.0,27.0,130.0
-8385,198.0,31.0,130.0
-11143,195.0,20.0,135.0
-11143,195.0,24.0,135.0
-15075,191.0,23.0,116.0
-15197,190.0,21.0,100.0
-17985,193.0,27.0,154.0
-22209,178.0,26.0,95.0
-23974,201.0,23.0,125.0
-27183,183.0,27.0,93.0
-29075,183.0,30.0,120.0
-37413,187.0,20.0,101.0
-37413,187.0,24.0,101.0
-38858,185.0,22.0,89.0
-38860,188.0,20.0,81.0
-38860,188.0,24.0,81.0
-40969,193.0,28.0,129.0
-44060,192.0,28.0,126.0
-46056,183.0,26.0,110.0
-48636,185.0,22.0,95.0
-48818,188.0,24.0,100.0
-64607,190.0,23.0,102.0
-65568,185.0,23.0,125.0
-67172,193.0,27.0,114.0
-71196,193.0,24.0,118.0
-72605,196.0,25.0,120.0
-73758,204.0,30.0,142.0
-75800,185.0,29.0,132.0
-76534,200.0,19.0,120.0
-85183,183.0,25.0,120.0
-85183,183.0,29.0,120.0
-85913,190.0,23.0,102.0
-86233,185.0,30.0,79.0
-87674,190.0,28.0,111.0
-88881,184.0,27.0,142.0
-96259,182.0,22.0,97.5
-102565,195.0,20.0,115.0
-102868,190.0,24.0,92.0
-109010,193.0,22.0,108.0
-109597,190.0,25.0,88.0
-111695,186.0,26.0,128.0
-115509,198.0,21.0,117.0
-125325,189.0,27.0,140.0
-128283,185.0,24.0,120.0
-131215,178.0,25.0,105.0
-131541,188.0,25.0,136.0
-131541,188.0,29.0,136.0
-4674,184.0,22.0,120.0
-5702,192.0,29.0,120.0
-10156,183.0,25.0,77.0
-10158,188.0,28.0,98.0
-11145,191.0,19.0,108.0
-11977,178.0,23.0,75.0
-12189,180.0,26.0,97.0
-13127,183.0,36.0,113.0
-20746,183.0,28.0,102.0
-24108,170.0,20.0,89.0
-26835,188.0,34.0,116.0
-35717,178.0,36.0,95.0
-35717,178.0,40.0,95.0
-43490,172.0,25.0,83.0
-47255,179.0,28.0,103.0
-59323,196.0,24.0,124.0
-61238,185.0,35.0,103.0
-64151,189.0,27.0,90.0
-64425,194.0,30.0,125.0
-70034,176.0,36.0,102.0
-70538,180.0,22.0,106.0
-70538,180.0,30.0,106.0
-71719,189.0,26.0,99.0
-77576,182.0,36.0,115.0
-83502,187.0,37.0,96.0
-84850,187.0,34.0,120.0
-85257,184.0,30.0,82.0
-85257,184.0,34.0,82.0
-86319,191.0,31.0,105.0
-87235,196.0,27.0,128.0
-87848,180.0,26.0,98.0
-91787,188.0,30.0,116.0
-96259,182.0,34.0,97.5
-102565,195.0,20.0,115.0
-103266,187.0,29.0,96.0
-103672,183.0,28.0,90.0
-103796,188.0,39.0,113.0
-105015,183.0,23.0,107.0
-110260,185.0,21.0,106.0
-110260,185.0,25.0,106.0
-110260,185.0,33.0,106.0
-111671,188.0,38.0,110.0
-115486,184.0,38.0,110.0
-118215,191.0,23.0,120.0
-120203,183.0,28.0,107.0
-120312,186.0,32.0,110.0
-120312,186.0,40.0,110.0
-121087,187.0,21.0,98.0
-126950,193.0,26.0,126.0
-134016,195.0,24.0,92.0
-135095,192.0,24.0,120.0
-135313,190.0,23.0,102.0
-135313,190.0,31.0,102.0
-2469,200.0,32.0,130.0
-2469,200.0,36.0,130.0
-3828,191.0,25.0,97.0
-6471,195.0,23.0,121.0
-16209,196.0,25.0,120.0
-18455,190.0,23.0,102.0
-22157,195.0,22.0,102.0
-22737,183.0,31.0,100.0
-22737,183.0,35.0,100.0
-23195,195.0,22.0,109.0
-25157,193.0,27.0,123.0
-25157,193.0,31.0,123.0
-25157,193.0,35.0,123.0
-25284,198.0,31.0,125.0
-30313,191.0,24.0,115.0
-30623,178.0,24.0,86.0
-38860,188.0,20.0,81.0
-39535,185.0,24.0,81.0
-39535,185.0,26.0,81.0
-40215,186.0,20.0,84.0
-41788,184.0,25.0,104.0
-46275,207.0,26.0,120.0
-50016,183.0,27.0,93.0
-50258,185.0,22.0,85.0
-52107,198.0,22.0,118.0
-57370,197.0,29.0,125.0
-63428,204.0,25.0,132.0
-65668,183.0,22.0,95.0
-69726,190.0,24.0,98.0
-77152,193.0,25.0,130.0
-77152,193.0,33.0,130.0
-82709,196.0,27.0,120.0
-86170,179.0,27.0,80.0
-86233,185.0,30.0,79.0
-87769,189.0,22.0,82.0
-88073,192.0,20.0,125.0
-96128,185.0,21.0,95.0
-96625,188.0,37.0,110.0
-98959,188.0,29.0,107.0
-107151,199.0,22.0,115.0
-109597,190.0,23.0,88.0
-109597,190.0,25.0,88.0
-123034,203.0,32.0,120.0
-130471,193.0,33.0,115.0
