--- logistic_regression/Training_/Y_test_HWAY.csv
@@ -1,136 +1,0 @@
-ID,MedalValue
-2201,0
-7587,0
-8238,0
-9786,0
-25346,0
-42303,0
-51071,0
-62308,0
-87674,0
-92739,0
-93433,0
-107613,0
-110212,0
-114187,0
-125791,0
-129478,0
-133843,0
-8761,0
-28219,0
-32764,0
-32957,0
-39294,0
-61718,0
-77425,0
-92409,0
-116206,0
-1856,0
-30559,0
-31616,0
-32297,0
-32800,0
-34006,0
-44591,0
-44597,0
-56250,0
-59781,0
-62194,0
-64396,0
-65694,0
-71588,0
-76347,0
-78659,0
-84515,0
-117099,0
-120053,0
-126502,0
-5159,1
-7927,1
-7968,1
-18655,1
-29666,1
-40969,1
-43949,1
-46317,1
-49109,1
-49817,1
-49910,1
-50258,1
-62308,1
-63392,1
-68647,1
-71196,1
-73758,1
-76534,1
-77472,1
-80287,1
-85913,1
-87674,1
-87674,1
-102565,1
-102565,1
-115290,1
-115992,1
-119961,1
-120385,1
-130080,1
-359,1
-5702,1
-13127,1
-22698,1
-27978,1
-35717,1
-46029,1
-60835,1
-61238,1
-63609,1
-63609,1
-77576,1
-77576,1
-83502,1
-87848,1
-95488,1
-100730,1
-104098,1
-113888,1
-118215,1
-128266,1
-135313,1
-2469,1
-15669,1
-17078,1
-25284,1
-27323,1
-28544,1
-30678,1
-41788,1
-44597,1
-46277,1
-50258,1
-53966,1
-54001,1
-57280,1
-57280,1
-61753,1
-64326,1
-79780,1
-88073,1
-88073,1
-88073,1
-96625,1
-100682,1
-100682,1
-102565,1
-107643,1
-107643,1
-109597,1
-110584,1
-117837,1
-117837,1
-118220,1
-121356,1
-123034,1
-129356,1
-130471,1
-131051,1
