--- logistic_regression/Training_/Y_train_HWAY.csv
@@ -1,253 +1,0 @@
-ID,MedalValue
-1695,0
-5191,0
-5803,0
-15458,0
-17923,0
-19770,0
-21249,0
-32496,0
-37878,0
-40323,0
-50016,0
-61166,0
-68411,0
-70432,0
-72065,0
-73758,0
-78020,0
-79025,0
-79660,0
-82976,0
-86380,0
-88479,0
-97119,0
-109597,0
-114240,0
-117837,0
-122671,0
-123168,0
-126095,0
-126492,0
-132654,0
-11145,0
-11446,0
-17103,0
-21546,0
-21629,0
-22698,0
-30152,0
-31090,0
-33773,0
-34592,0
-37085,0
-39294,0
-40373,0
-40373,0
-41106,0
-46953,0
-52324,0
-55996,0
-59290,0
-61558,0
-62267,0
-70341,0
-83502,0
-95965,0
-101611,0
-103116,0
-104098,0
-105132,0
-110208,0
-117537,0
-121464,0
-123497,0
-129419,0
-129419,0
-2198,0
-2469,0
-9379,0
-11492,0
-12717,0
-15754,0
-21571,0
-37262,0
-38731,0
-38858,0
-40373,0
-42127,0
-50076,0
-53966,0
-56182,0
-61225,0
-61225,0
-64687,0
-70034,0
-73969,0
-75708,0
-75708,0
-76347,0
-79979,0
-83256,0
-83721,0
-83721,0
-84515,0
-89010,0
-96625,0
-97430,0
-108505,0
-108528,0
-112391,0
-112845,0
-117225,0
-117443,0
-120792,0
-128842,0
-131821,0
-4094,1
-6810,1
-7968,1
-8385,1
-8385,1
-11143,1
-11143,1
-15075,1
-15197,1
-17985,1
-22209,1
-23974,1
-27183,1
-29075,1
-37413,1
-37413,1
-38858,1
-38860,1
-38860,1
-40969,1
-44060,1
-46056,1
-48636,1
-48818,1
-64607,1
-65568,1
-67172,1
-71196,1
-72605,1
-73758,1
-75800,1
-76534,1
-85183,1
-85183,1
-85913,1
-86233,1
-87674,1
-88881,1
-96259,1
-102565,1
-102868,1
-109010,1
-109597,1
-111695,1
-115509,1
-125325,1
-128283,1
-131215,1
-131541,1
-131541,1
-4674,1
-5702,1
-10156,1
-10158,1
-11145,1
-11977,1
-12189,1
-13127,1
-20746,1
-24108,1
-26835,1
-35717,1
-35717,1
-43490,1
-47255,1
-59323,1
-61238,1
-64151,1
-64425,1
-70034,1
-70538,1
-70538,1
-71719,1
-77576,1
-83502,1
-84850,1
-85257,1
-85257,1
-86319,1
-87235,1
-87848,1
-91787,1
-96259,1
-102565,1
-103266,1
-103672,1
-103796,1
-105015,1
-110260,1
-110260,1
-110260,1
-111671,1
-115486,1
-118215,1
-120203,1
-120312,1
-120312,1
-121087,1
-126950,1
-134016,1
-135095,1
-135313,1
-135313,1
-2469,1
-2469,1
-3828,1
-6471,1
-16209,1
-18455,1
-22157,1
-22737,1
-22737,1
-23195,1
-25157,1
-25157,1
-25157,1
-25284,1
-30313,1
-30623,1
-38860,1
-39535,1
-39535,1
-40215,1
-41788,1
-46275,1
-50016,1
-50258,1
-52107,1
-57370,1
-63428,1
-65668,1
-69726,1
-77152,1
-77152,1
-82709,1
-86170,1
-86233,1
-87769,1
-88073,1
-96128,1
-96625,1
-98959,1
-107151,1
-109597,1
-109597,1
-123034,1
-130471,1
